refactor(server): move mpe controller specific tools into the controller file
